NET-2: moved packet to phase1packet.go
